requirements.txt
attrs==19.1.0
falcon==2.0.0
falcon-autocrud==1.0.36
gunicorn==20.0.4
jsonschema==3.0.1
marshmallow==2.19.5
psycopg2==2.8.3
pyrsistent==0.15.4
python-dateutil==2.8.0
six==1.12.0
SQLAlchemy==1.3.6
openpyxl==3.0.3
meinheld==1.0.1
python-rapidjson==0.9.1
webargs==6.0.0